Fix: Handle errors caused by presence of charset in blob types
Before the change, a blob type containing a MIME type with the `charset` set
could cause a breaking error. This commit addresses the issue.

It might be an imperfect solution since there might be other charsets, and the
`blob.text()` function always assumes UTF-8. 
However, the alternative would be usage of the `FileReader.readAsText()` which will require additional changes, as the
latter uses callbacks instead of promises, and it might be deprecated later due to this.

The possibility of using another encoding is negligible. Moreover, support for
it might be introduced later in case someone reports a problem.